community(feat): further componentized events, moved events to state due to utilization of condditional rendering
