+++ src/components/state/Events/PastEvents/PastEvents.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import { EventDetails, EventLabel, EventName, AboutHeader, AboutDesc, EventImage } from '../EventDetails';
+import { EventDetails, EventLabel, EventName, AboutHeader, AboutDesc, EventImage } from '../EventDetails/EventDetails';
 import RedirectButton from '../../../state/RedirectButton/RedirectButton';
 
 class PastEvents extends Component {
